Remove length limit from displayedValue, and move useMainValue.js to the proper hooks/ folder.
