+++ packages/core/admin/admin/src/content-manager/components/DynamicZone/components/Component/hooks/tests/getDisplayedValue.test.js
